Use 2.0-style mode to remove cascade_backrefs
SQLAlchemy has long had a behavior of cascading objects into the Session
based on backref assignment. This behavior and has been the cause of
session instance conflicts and thus has not been convenient for us. To
remedy this, we use the SQLAlchemy 2.0-style mode which disables this
behavior by default.

For more information, see the following:

    http://www.sqlalchemy.org/trac/ticket/5150